+++ src/lib.js
@@ -1,0 +1,9 @@
+// During development, using ESM for module resolution requires that
+// all paths must be explicit – so a path to './lib' will not automatically
+// find the index.js file. So we use lib.js instead of lib/index.js.
+export * as constants from "./lib/constants.js";
+export * as utils from "./lib/utils.js";
+export { getGlobalZjax } from "./lib/global.js";
+export { debug } from "./lib/debug.js";
+export { parseZSwaps } from "./lib/zswaps.js";
+export { parseZActions } from "./lib/zactions.js";