+++ src/lib/debug.js
@@ -1,5 +1,5 @@
 export function debug() {
-  if (!window.zjax.debug) {
+  if (!(window.zjax && window.zjax.debug)) {
     return;
   }
   console.log("ZJAX DEBUG:", ...arguments);